split-out opentype/fvar code-tests
